Refactor fetchUser action creator so that redux-thunk can capture it
redux-thunk is a middleware that inspects the value returned by an
action creator. If it detects that the action creator returns a
function instead of an action, it will call the function passing to
it a dispatch function as an argument. Once code inside the
function is executed, the dispatch function must be called passing
to it the action to be forwared to the reducers.