Add action creator 'handleToken'
the 'token' parameter is the token that we get back from the
stripe api.

the function sends a post request to update the user's credits to
the backend server.  The function then dispatches the action with
action type 'FETCH_USER' to update the user model inside of the
auth reducer.